refactor: introduce repository pattern, service layer, and schema package
Routes now follow a strict three-tier call chain:
  route → service → repository

- repositories/: all SQLAlchemy queries with exception translation
  (IntegrityError → ConflictError, SQLAlchemyError → RepositoryError)
- services/pet.py: business logic layer between routes and the repository
- schemas/: Pydantic request/response models extracted from route files
- exceptions.py: typed domain exception hierarchy
- api/exception_handlers.py: maps domain exceptions to HTTP status codes
  (NotFoundError → 404, ConflictError → 409, RepositoryError → 500)
- crud/ shims re-export from repositories/ for backwards compatibility
  with main.py, webhook.py, mcp/server.py, and existing tests
