+++ src/github_tamagotchi/repositories/__init__.py
@@ -1,0 +1,1 @@
+"""Data access layer. All SQLAlchemy queries live here."""
