+++ src/github_tamagotchi/schemas/__init__.py
@@ -1,0 +1,1 @@
+"""Request/response schemas (Pydantic DTOs) for the API."""
